requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
groq==0.9.0
PyMuPDF==1.24.5
python-docx==1.1.2
streamlit==1.36.0
requests==2.32.3
pydantic==2.7.1
